Added the bing_linkedin_cache module.
Reactored the github modules and supporting framework code to fix several bugs and improve code quality. Resolves issue #204.

Made some functional changes to the github modules.

Made some aesthetic changes to the github_dorks and gists_search modules.

Fixed a bug in the GoogleWebMixin that broke CAPTCHA bypass functionality. Resolves issue #191. (@jnqpblc)

Refactored the interactive output system. The add_* methods now handle output based on the uniqueness of the record being added to the database. This reduces the burden on the module developer of having to handle terminal output and provides a more consistent user experience.

Replaced the separate DEBUG and VERBOSE global options with a single VERBOSITY option.